feat(C++): Add `delete` vs `delete[]`
First, let's have a class `A` which prints something in the constructor, and something in the destructor.
When we `new A[42]`, we see the constructor print 42 times. So far so normal.

But how is it that when we pass that memory address around and eventually `delete[]` it, the program knows at runtime **how many destructors have to be called**?

## Most modern compilers do a secret trick when we `new[N]`

They not only allocate memory for `N` objects, but also for 1 extra `size_t`.
That `size_t` stores the amount of objects that follow, and is stored at the very beginning of the block of memory that gets allocated.
And with that `size_t` the program knows how many destructors to call.

## But why does it crash when we `delete` a pointer that was `new[]`ed?

The pointer that `new[]` returns though starts from the first actual object, so not including the `size_t`. That's important so we can do `array[0]` to access the first element.

But let's keep in mind that the start of the actually allocated memory block starts 1 `size_t` before the pointer that we got!
That is why we have the `delete[]`.
The main difference between `delete[]` and `delete` is that `delete[]` frees the memory 1 `size_t` before the address that we give it, while `delete` frees exactly the address we give it.

That explains why we get an invalid free when we pass a pointer that was `new[]`ed to `delete` instead of `delete[]`.
`delete[]` assumes the actual memory address to free is at an offset, while `delete` does not.

It is exactly the same thing as doing this in C:
```c
char *ptr = malloc(10);
ptr++;
free(ptr);
```

---

This is called "over-allocation", and it is a common technique to implement all sorts of memory tracking mechanisms, like reference-counting.

---

Source:
https://isocpp.org/wiki/faq/compiler-dependencies#num-elems-in-new-array-overalloc